requirements.txt
docutils
pillow

django-dept-directory
django-webcal
django-markuphelpers
